requirement.txt
streamlit==1.35.0
pandas==2.2.2
jinja2==3.1.3
weasyprint==60.2
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
num2words==0.5.13
